Add GraphQL configuration for handling server queries and mutations
This commit adds GraphQL server configuration and new functionalities.
The changes include adding new `go.sum` file including all the modules and imports required for the server such as '99designs/gqlgen' and creating a `gqlgen.yml` file used to configure code generation for GraphQL.
`server.go` is added to handle http requests and serves a GraphQL playground on the root path.
`schema.graphqls` is used to define a schema with a simple Todo and User types, and a mutation to create new todos.
The `graph/generated.go` is generated based on the schema and is used by the gqlgen library to match the schema to the resolver functions.